Modify: ユーザーに関するVuexStoreをモジュール化
