+++ app/javascript/router/index.js
@@ -37,7 +37,7 @@ const router = new VueRouter({
 })
 
 router.beforeEach((to, _from, next) =>{
-  store.dispatch('fetchAuthUser').then((authUser) => {
+  store.dispatch('users/fetchAuthUser').then((authUser) => {
     if (to.matched.some(record => record.meta.requireAuth) && !authUser) {
       next({ name: 'LoginIndex' });
     } else {
